# libs/shared/openapi-spec/src/main/resources/openapi/bundled-synaptiq-v1.yaml
openapi: 3.1.0
info:
  title: Synaptiq API
  description: |
    AI-Powered Conversational Commerce Platform — REST API

    Synaptiq enables businesses to deploy intelligent, context-aware AI shopping assistants
    with full catalog management, conversation orchestration, workflow automation,
    and multi-tenant governance. This specification covers all bounded contexts:
    Authentication, Tenants, Catalog, Chat, Actions, Workflows, Analytics,
    Branding, Tenant Configuration, and Schema Registry.
  version: 1.0.0
  contact:
    name: Synaptiq Team
    url: https://synaptiq.dev
  license:
    name: Proprietary
    url: https://synaptiq.dev/license
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.synaptiq.dev
    description: Production server
...
paths:
...
  /api/v1/applications:
    get:
      operationId: listApplications
      summary: List applications for a tenant
      tags:
        - Applications
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Application list
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApplicationResponse'
    post:
      operationId: createApplication
      summary: Create a new application
      tags:
        - Applications
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        '201':
          description: Application created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
  /api/v1/applications/{appId}:
    get:
      operationId: getApplication
      summary: Get application by ID
      tags:
        - Applications
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
    delete:
      operationId: deleteApplication
      summary: Delete an application
      tags:
        - Applications
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - name: appId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Application deleted
...
  /api/v1/config/branding:
    get:
      operationId: getBranding
      summary: Get branding config
      tags:
        - Branding
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Branding config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingResponse'
    patch:
      operationId: updateBranding
      summary: Update branding
      tags:
        - Branding
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandingResponse'
      responses:
        '200':
          description: Branding updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingResponse'
  /api/v1/config/branding/public:
    get:
      summary: Get Public Branding
      description: Retrieve public branding information for a tenant without authentication.
      operationId: getPublicBranding
      tags:
        - Branding
      parameters:
        - name: X-Tenant-ID
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public branding configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandingResponse'
        '404':
          description: Tenant not found
        '500':
          description: Internal server error
  /api/v1/config/themes:
    get:
      operationId: getThemes
      summary: Get theme presets
      tags:
        - Branding
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Theme presets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThemeListResponse'
  /api/v1/config/contrast:
    get:
      operationId: checkContrast
      summary: Check color contrast ratio
      tags:
        - Branding
      parameters:
        - name: fg
          in: query
          required: true
          schema:
            type: string
        - name: bg
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Contrast result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContrastCheckResponse'
  /api/v1/config/personalization:
    get:
      operationId: getPersonalization
      summary: Get personalization settings
      tags:
        - Branding
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      responses:
        '200':
          description: Personalization settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalizationResponse'
    patch:
      operationId: updatePersonalization
      summary: Update personalization
      tags:
        - Branding
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonalizationResponse'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalizationResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: The tenant identifier for multi-tenant isolation
      schema:
        type: string
        minLength: 1
        maxLength: 100
  schemas:
...
    ApplicationResponse:
      type: object
      description: Full application configuration response
      properties:
        appId:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        isDefault:
          type: boolean
        status:
          $ref: '#/components/schemas/ApplicationStatus'
        aiPersona:
          $ref: '#/components/schemas/AiPersonaResponse'
        guardrails:
          $ref: '#/components/schemas/GuardrailsResponse'
        branding:
          $ref: '#/components/schemas/BrandingResponse'
        components:
          $ref: '#/components/schemas/ComponentsToggleResponse'
        actions:
          $ref: '#/components/schemas/ActionsConfigResponse'
        personalization:
          $ref: '#/components/schemas/PersonalizationResponse'
        themes:
          type: array
          items:
            $ref: '#/components/schemas/ThemePresetResponse'
        dataSourceIds:
          type: array
          items:
            type: string
        llmOverride:
          $ref: '#/components/schemas/LlmProviderResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateApplicationRequest:
      type: object
      description: Create a new application
      required:
        - name
        - slug
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        icon:
          type: string
        isDefault:
          type: boolean
          default: false
        aiPersona:
          $ref: '#/components/schemas/AiPersonaResponse'
        guardrails:
          $ref: '#/components/schemas/GuardrailsResponse'
        branding:
          $ref: '#/components/schemas/BrandingResponse'
        components:
          $ref: '#/components/schemas/ComponentsToggleResponse'
        actions:
          $ref: '#/components/schemas/ActionsConfigResponse'
        personalization:
          $ref: '#/components/schemas/PersonalizationResponse'
        themes:
          type: array
          items:
            $ref: '#/components/schemas/ThemePresetResponse'
        dataSourceIds:
          type: array
          items:
            type: string
        llmOverride:
          $ref: '#/components/schemas/LlmProviderResponse'
    ApplicationStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - PAUSED
        - ARCHIVED
...
    BrandingResponse:
      type: object
      properties:
        logoUrl:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        backgroundStyle:
          type: string
        headingFont:
          type: string
        bodyFont:
          type: string
        faviconUrl:
          type: string
        pageTitle:
          type: string
        showPlatformBranding:
          type: boolean
    ThemePresetResponse:
      type: object
      properties:
        themeId:
          type: string
        name:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        backgroundStyle:
          type: string
        isDefault:
          type: boolean
    ThemeListResponse:
      type: object
      properties:
        themes:
          type: array
          items:
            $ref: '#/components/schemas/ThemePresetResponse'
    ContrastCheckResponse:
      type: object
      properties:
        foreground:
          type: string
        background:
          type: string
        ratio:
          type: number
          format: double
        aaNormal:
          type: boolean
        aaLarge:
          type: boolean
        aaaNormal:
          type: boolean
    PersonalizationResponse:
      type: object
      properties:
        allowThemeSwitch:
          type: boolean
        allowFontSwitch:
          type: boolean
        allowBubbleStyle:
          type: boolean
    AiPersonaResponse:
      type: object
      properties:
        displayName:
          type: string
        tone:
          type: string
        customInstruction:
          type: string
        welcomeMessage:
          type: string
        starterPrompts:
          type: array
          items:
            type: string
    GuardrailsResponse:
      type: object
      properties:
        outOfScopeMessage:
          type: string
        recommendationMode:
          type: boolean
        language:
          type: string
    LlmProviderResponse:
      type: object
      properties:
        provider:
          type: string
        modelId:
          type: string
        isByok:
          type: boolean
    ComponentsToggleResponse:
      type: object
      properties:
        itemCard:
          type: boolean
        itemGrid:
          type: boolean
        itemDetail:
          type: boolean
        comparisonTable:
          type: boolean
        filterSummary:
          type: boolean
        resultCount:
          type: boolean
        emptyState:
          type: boolean
        actionConfirm:
          type: boolean
        infoBanner:
          type: boolean
    ActionsConfigResponse:
      type: object
      properties:
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ActionEntryResponse'
        enquiryWebhookUrl:
          type: string
        enquiryEmail:
          type: string
    ActionEntryResponse:
      type: object
      properties:
        actionId:
          type: string
        enabled:
          type: boolean
        label:
          type: string
...